Changed to decimal to fix floating point errors and added a filter to buildings. Also added storage amounts and some sorting.

+++ DuneData/Model/Item.cs
@@ -3,11 +3,11 @@
     public class Item
     {
         public Guid Id { get; set; } = Guid.NewGuid();
-        public double Volume { get; set; }
+        public decimal Volume { get; set; }
         public string Name { get; set; }
         public ItemType Type { get; set; }
         public string Image { get; set; } = string.Empty;
-        public Item(string name, double vol, ItemType type, string image = "")
+        public Item(string name, decimal vol, ItemType type, string image = "")
         {
             Name = name;
             Volume = vol;
